+++ packages/commcare/src/Adaptor.js
@@ -4,7 +4,7 @@ import { Blob } from 'node:buffer';
 import js2xmlparser from 'js2xmlparser';
 import xlsx from 'xlsx';
 
-import { request, prepareNextState } from './Utils';
+import * as util from './Utils';
 /**
  * State object
  * @typedef {Object} CommcareHttpState
@@ -82,7 +82,7 @@ export function get(path, params = {}, callback = s => s) {
 
       do {
         // Make the first request
-        const response = await request(
+        const response = await util.request(
           state.configuration,
           `/a/${domain}/api/v0.5/${resolvedPath}`,
           {
@@ -92,7 +92,7 @@ export function get(path, params = {}, callback = s => s) {
           }
         );
 
-        nextState = prepareNextState(state, response, callback);
+        nextState = util.prepareNextState(state, response, callback);
         // If the server tells us there's another page of data, setup
         // the next request to get it
         if (response?.body?.meta?.next) {
@@ -156,7 +156,7 @@ export function post(path, data, params = {}, callback = s => s) {
     );
 
     try {
-      const response = await request(
+      const response = await util.request(
         state.configuration,
         `/a/${domain}/api/v0.5/${resolvedPath}`,
         {
@@ -167,7 +167,7 @@ export function post(path, data, params = {}, callback = s => s) {
         }
       );
 
-      return prepareNextState(state, response, callback);
+      return util.prepareNextState(state, response, callback);
     } catch (e) {
       throw e.body.error ?? e;
     }
@@ -222,12 +222,12 @@ export function submitXls(data, params) {
     form.append('create_new_cases', create_new_cases);
 
     try {
-      const response = await request(state.configuration, path, {
+      const response = await util.request(state.configuration, path, {
         method: 'POST',
         data: form,
       });
 
-      return prepareNextState(state, response);
+      return util.prepareNextState(state, response);
     } catch (e) {
       throw e.body ?? e;
     }
@@ -265,14 +265,14 @@ export function submit(data) {
     console.log('Raw JSON body: '.concat(JSON.stringify(jsonBody)));
     console.log('X-form submission: '.concat(body));
 
-    const response = await request(state.configuration, path, {
+    const response = await util.request(state.configuration, path, {
       method: 'POST',
       data: body,
       contentType: 'text/xml',
       parseAs: 'text',
     });
 
-    return prepareNextState(state, response);
+    return util.prepareNextState(state, response);
   };
 }
 
@@ -298,12 +298,12 @@ export function fetchReportData(reportId, params, postUrl) {
 
     console.log('with params: '.concat(JSON.stringify(params)));
 
-    const { body: reportData } = await request(state.configuration, path, {
+    const { body: reportData } = await util.request(state.configuration, path, {
       method: 'GET',
       contentType: 'application/json',
     });
 
-    await request(state.configuration, postUrl, {
+    await util.request(state.configuration, postUrl, {
       method: 'POST',
       params,
       data: reportData,
@@ -316,6 +316,133 @@ export function fetchReportData(reportId, params, postUrl) {
   };
 }
 
+/**
+ * Make a general HTTP request against the Commcare server.
+ * @example <caption>Make a GET request to get cases</caption>
+ * request("GET", "/a/asri/api/v0.5/case");
+ * @function
+ * @public
+ * @param {string} method - HTTP method to use
+ * @param {string} path - Path to resource
+ * @param {object} body - Object which will be attached to the body
+ * @param {RequestOptions} options - Optional request params
+ * @returns {Operation}
+ * @state {CommcareHttpState}
+ */
+export function request(method, path, body, options = {}) {
+  return async state => {
+    const [resolvedMethod, resolvedPath, resolvedBody, resolvedOptions] =
+      expandReferences(state, method, path, body, options);
+      const response = await util.request(state.configuration, resolvedPath, {
+        method: resolvedMethod,
+        data: resolvedBody,
+        params: resolvedOptions,
+        contentType: 'application/json',
+      });
+
+      return util.prepareNextState(state, response);
+
+  };
+}
+
+/**
+ * Bulk upload data to CommCare for case-data or lookup-table. Accepts an array of objects, converts them into
+ * an XLS representation, and uploads.
+ * @public
+ * @function
+ * @example <caption>Upload a single row of data for case-data</caption>
+ * bulk(
+ *    [
+ *      {name: 'Mamadou', phone: '000000'},
+ *    ],
+ *    {
+ *      case_type: 'student',
+ *      search_field: 'external_id',
+ *      create_new_cases: 'on',
+ *    }
+ * )
+ * @example <caption>Upload a single row of data for a lookup-table</caption>
+ * bulk(
+ *     'lookup-table'
+ *  {
+ *    types: [{
+ *
+ *   'DELETE(Y/N)':'N',
+ *   table_id: 'fruit',
+ *   'is_global?':'yes',
+ *   'field 1': 'type',
+ *   'field 2': 'name',
+ *      }],
+ *      fruit: [{
+ *       UID: '',
+ *        'DELETE(Y/N)':'N',
+ *       'field:type': 'citrus',
+ *        'field:name': 'Orange',
+ *     }],
+ *   }
+ * )
+ * @param {string} type - case-data or lookup-table
+ * @param {array} data - Array of objects to upload
+ * @param {Object} params - Input parameters, see {@link https://dimagi.atlassian.net/wiki/spaces/commcarepublic/pages/2143946459/Bulk+Upload+Case+Data CommCare docs} for case-data and {@link https://dimagi.atlassian.net/wiki/spaces/commcarepublic/pages/2143946023/Bulk+upload+Lookup+Tables Commcare Docs} for lookup-table.
+ * @state data - the response from the CommCare Server
+ * @returns {Operation}
+ */
+export function bulk(type, data, params) {
+  return async state => {
+    const { domain } = state.configuration;
+
+    const [json] = expandReferences(state, data);
+    let path, file;
+
+    const workbook = xlsx.utils.book_new();
+
+    if (type.toLowerCase() === 'lookup-table') {
+      path = `/a/${domain}/fixtures/fixapi/`;
+      file = 'file-to-upload';
+      // append types and lookup-table name xlsx
+      Object.keys(json).forEach(sectionName => {
+        const sectionData = data[sectionName];
+
+        const newSheet = xlsx.utils.json_to_sheet(sectionData);
+        xlsx.utils.book_append_sheet(workbook, newSheet, sectionName);
+      });
+    } else if (type.toLowerCase() === 'case-data') {
+      path = `/a/${domain}/importer/excel/bulk_upload_api/`;
+      file = 'file';
+      const worksheet = xlsx.utils.json_to_sheet(json);
+      const ws_name = 'SheetJS';
+      xlsx.utils.book_append_sheet(workbook, worksheet, ws_name);
+    } else {
+      const e = new Error('Unrecognized type');
+      e.description = `The type key was not recognized: ${type}`;
+      e.fix = 'Set type to case-data or lookup-table';
+      throw e;
+    }
+
+    const buffer = xlsx.write(workbook, { type: 'buffer', bookType: 'xlsx' });
+
+    const form = new FormData();
+
+    form.append(file, new Blob([buffer]), 'data.xlsx');
+
+    for (const key in params) {
+      form.append(key, params[key]);
+    }
+
+    const response = await util.request(state.configuration, path, {
+      method: 'POST',
+      data: form,
+    });
+
+    return util.prepareNextState(state, {
+      ...response,
+      body: {
+        message: response.body.message.replace(/\n/g, ''),
+        code: response.body.code,
+      },
+    });
+  };
+}
 export {
   fn,
   fnIf,